Move capitalizeFirstLetter to utils folder
