+++ client/components/resource-details/SubmitReview.tsx
@@ -11,8 +11,7 @@ import {
 import Modal from 'react-modal'
 import { useState, Fragment } from 'react'
 import ReactStars from 'react-stars'
-import { Faculty, Status, resourceLookup } from '../../utils'
-import { capitalizeFirstLetter } from '../common/Utils'
+import { Faculty, Status, resourceLookup, capitalizeFirstLetter } from '../../utils'
 
 const SubmitButton = styled.button`
   font-size: 18px;
