+++ client/utils/functions.ts
