+++ client/utils/index.ts
@@ -1,4 +1,5 @@
 export { indexData } from './IndexUtils'
 // export { deserializeClubsAndEvents } from './serde'
-export type { Id, Resource, House, StudySpot, Washroom } from './types'
+export type { Id, Resource, House, StudySpot, Washroom, Review } from './types'
 export { Faculty, Status, resourceLookup } from './types'
+export { capitalizeFirstLetter } from './functions'
